ImageProcessor/requirements.txt
pytesseract
Pillow
imagehash
